+++ packages/aws-client/src/services.ts
@@ -1,4 +1,4 @@
-import { ApiGatewayV2Client } from '@aws-sdk/client-apigatewayv2';
+import { APIGatewayClient } from '@aws-sdk/client-api-gateway';
 import { DynamoDB } from '@aws-sdk/client-dynamodb';
 import { DynamoDBStreamsClient } from '@aws-sdk/client-dynamodb-streams';
 import { S3 } from '@aws-sdk/client-s3';
@@ -7,7 +7,7 @@ import * as AWS from 'aws-sdk';
 
 import awsClient from './client';
 
-export const apigateway = awsClient(ApiGatewayV2Client, '2015-07-09');
+export const apigateway = awsClient(APIGatewayClient, '2015-07-09');
 export const ecs = awsClient(AWS.ECS, '2014-11-13');
 export const ec2 = awsClient(AWS.EC2, '2016-11-15');
 export const s3 = awsClient(S3, '2006-03-01');